refactor(cypress): polish example components
