+++ packages/e2e-cypress/src/templates/cct/base/src/components/QuasarButton.vue
@@ -3,10 +3,9 @@
 </template>
 
 <script>
-export default {
+import { defineComponent } from 'vue';
+
+export default defineComponent({
   name: 'QuasarButton',
-  setup() {
-    return {};
-  },
-};
+});
 </script>
