+++ packages/e2e-cypress/src/templates/cct/base/src/components/QuasarPageSticky.vue
@@ -7,13 +7,12 @@
 </template>
 
 <script>
-export default {
+import { defineComponent } from 'vue';
+
+export default defineComponent({
   name: 'QuasarPageSticky',
   props: {
     title: String,
   },
-  setup() {
-    return {};
-  },
-};
+});
 </script>
