+++ packages/e2e-cypress/src/templates/cct/examples-js/src/components/___tests__/QuasarButton.spec.js
@@ -1,5 +1,5 @@
 import { mount } from '@cypress/vue';
-import QuasarButton from './../QuasarButton.vue';
+import QuasarButton from '../QuasarButton.vue';
 
 describe('QuasarButton', () => {
   it('renders a message', () => {
