+++ packages/e2e-cypress/src/templates/cct/examples-ts/src/components/___tests__/QuasarDialog.spec.ts
@@ -1,6 +1,6 @@
 import { mount } from '@cypress/vue';
 import DialogWrapper from 'app/test/cypress/wrappers/DialogWrapper.vue';
-import QuasarDialog from './../QuasarDialog.vue';
+import QuasarDialog from '../QuasarDialog.vue';
 
 describe('QuasarDialog', () => {
   it('should show a dialog', () => {
