+++ packages/e2e-cypress/src/templates/cct/examples-ts/src/components/___tests__/QuasarDrawer.spec.ts
@@ -1,6 +1,6 @@
 import { mount } from '@cypress/vue';
 import LayoutContainer from 'app/test/cypress/wrappers/LayoutContainer.vue';
-import QuasarDrawer from './../QuasarDrawer.vue';
+import QuasarDrawer from '../QuasarDrawer.vue';
 
 describe('QuasarDrawer', () => {
   it('should show a drawer', () => {
