+++ packages/e2e-cypress/src/templates/cct/examples-ts/src/components/___tests__/QuasarPageSticky.spec.ts
@@ -1,6 +1,6 @@
 import { mount } from '@cypress/vue';
 import LayoutContainer from 'app/test/cypress/wrappers/LayoutContainer.vue';
-import QuasarPageSticky from './../QuasarPageSticky.vue';
+import QuasarPageSticky from '../QuasarPageSticky.vue';
 
 describe('QuasarPageSticky', () => {
   it('should show a drawer', () => {
