+++ packages/e2e-cypress/src/templates/cct/examples-ts/src/components/___tests__/QuasarTooltip.spec.ts
@@ -1,5 +1,5 @@
 import { mount } from '@cypress/vue';
-import QuasarTooltip from './../QuasarTooltip.vue';
+import QuasarTooltip from '../QuasarTooltip.vue';
 
 describe('QuasarTooltip', () => {
   it('should show a tooltip', () => {
